% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{check_cran_results}
\alias{check_cran_results}
\title{Deprecated functions}
\usage{
check_cran_results(...)
}
\arguments{
\item{...}{see documentation for \code{cran_results} and \code{summary_cran_details}}
}
\description{
Deprecated functions provided for back compatibility.
}
