% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn.R
\name{cran_check_results}
\alias{cran_check_results}
\title{Table of CRAN check results}
\usage{
cran_check_results(email = NULL, pkg = NULL, show = c("error", "fail",
  "warn", "note", "ok"))
}
\arguments{
\item{email}{email address for package maintainers (character
vector)}

\item{pkg}{package names (character vector)}

\item{show}{columns of the data frame to show (all are shown by default)}
}
\value{
a data frame that tabulates the number of CRAN platforms
    that return errors, warnings, notes, or OK for the packages.
}
\description{
Make a table that summarizes the results of the CRAN checks for a
set of packages specified by a maintainer or by names.
}
\details{
Given the email address of a package maintainer, and/or a vector
of package names, returns a tibble that allows you to detect
potential issues with your packages on CRAN.
}

