% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn-package.R
\docType{package}
\name{foghorn}
\alias{foghorn}
\alias{foghorn-package}
\title{CRAN check results at the terminal}
\description{
\code{foghorn} brings the results of the CRAN checks into your R
terminal. The results of the checks can be summarized as a
tibble, or as a summary.
}
\details{
The package also provides functions to view the logs of the CRAN
check results, or to visit the web pages hosted by CRAN.
}

