% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn.R
\name{summary_cran_checks}
\alias{summary_cran_checks}
\title{Summary of the CRAN check results}
\usage{
summary_cran_checks(email = whoami::email_address(), package = NULL)
}
\arguments{
\item{email}{email address for package maintainers (character
vector)}

\item{package}{package names (character vector)}
}
\value{
Prints the packages that return errors, warnings, and
    notes on the CRAN platforms. The number in parenthesis after
    the name of the packages indicates the number of CRAN
    platforms that produce these results.
}
\description{
Given the email address of a package maintainer, and/or a vector
of package names, it displays at the console a summary of the
check results run on the CRAN platforms. This function is designed
to be included in your .Rprofile to be run (periodically) at start
up.
}
\details{
This package tries to guess your email address by using the
\code{\link[whoami]{email_address}} function. You can specify email
addresses manually.
}
\examples{
\dontrun{ summary_cran_checks(email =
    c("user1@company1.com", "user2@company2.com"))
    summary_cran_checks(email = "user1@company1.com", package =
    c("pkg1", "pkg2")) }
}

