% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn.R
\name{visit_cran_check}
\alias{visit_cran_check}
\title{Visit the CRAN check results page for a package}
\usage{
visit_cran_check(pkg)
}
\arguments{
\item{pkg}{name of the package to check the results for}
}
\value{
The URL from the CRAN check results page invisibly
}
\description{
Visit the page in your web browser for a given package.
}

