% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn-package.R
\docType{package}
\name{foghorn}
\alias{foghorn}
\alias{foghorn-package}
\title{CRAN check results at the terminal}
\description{
\code{foghorn} brings the results of the CRAN checks into your R
terminal. This package is intended for developers that have
published packages on CRAN. The results of the checks can be
summarized as a tibble (see \code{\link{check_cran_results}}), or
as \dQuote{pretty} terminal output (see
\code{\link{summary_cran_results}}).  The package also provides
functions to view the logs of the CRAN check results (see
\code{\link{show_cran_results}}), or to visit the web pages hosted
by CRAN (see \code{\link{visit_cran_check}}).
}

