% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foghorn.R
\name{show_cran_results}
\alias{show_cran_results}
\title{Show the CRAN check results for a package}
\usage{
show_cran_results(pkg, show_log = TRUE)
}
\arguments{
\item{pkg}{name of the package on CRAN}

\item{show_log}{Should the messages of the \dQuote{Check Details}
be printed? (logical)}
}
\value{
\code{NULL}, used for its side effect of printing the CRAN
    messages
}
\description{
Given a package name published on CRAN, print the outcomes of
checks that return notes, warnings or errors, and optionally print
the messages.
}

